+++ R/yaml.R
@@ -115,11 +115,12 @@ yaml_handlers = function(h, envir) {
 #' Split a document into the YAML metadata (which starts with `---` in the
 #' beginning of the document) and the body. The YAML metadata will be parsed.
 #' @param x A character vector of the document content.
+#' @param ... Arguments to be passed to `yaml_load()`.
 #' @export
 #' @return A list of components `yaml` and `body`.
 #' @examples
 #' xfun::yaml_body(c('---', 'title: Hello', 'output: markdown::html_document', '---', '', 'Content.'))
-yaml_body = function(x) {
+yaml_body = function(x, ...) {
   i = grep('^---\\s*$', x)
   n = length(x)
   res = if (n < 2 || length(i) < 2 || (i[1] > 1 && !all(is_blank(x[seq(i[1] - 1)])))) {
@@ -128,7 +129,7 @@ yaml_body = function(x) {
     yaml = x[i[1]:i[2]], body = c(rep('', i[2]), tail(x, n - i[2]))
   )
   if ((n <- length(res$yaml)) >= 3) {
-    res$yaml = yaml_load(res$yaml[-c(1, n)])
+    res$yaml = yaml_load(res$yaml[-c(1, n)], ...)
   }
   res
 }